backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
# Pinned stack (Supabase + websockets); see note on realtime below.
supabase==2.11.0
websockets==13.1
python-multipart==0.0.12
# realtime==0.0.3 CANNOT be installed with supabase==2.11.0 (supabase requires realtime>=2.0.0,<3.0.0).
# Use realtime 2.x so imports match websockets 13.x and supabase’s realtime client.
realtime>=2.0.0,<3.0.0
groq>=0.11.0
pydantic>=2.9.0
pytesseract>=0.3.10
Pillow>=10.0.0
